backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
package-mode = false
description = "Employee attendance & team supervision API"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "shortuuid (>=1.0.13,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
ruff = "^0.14.0"
isort = "^6.1.0"
pre-commit = "^4.3.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
coverage = "^7.10.7"
factory-boy = "^3.3.3"
freezegun = "^1.5.5"

